+++ packages/mwwm/CHANGELOG.md
@@ -1,4 +1,9 @@
+[0.1.0+1] 
 
+* [BREAKING CHANGE!] Change API of Performer and Model. Now R may be any object.
+* Add special performers for useful cases: FuturePerformer and StreamPerformer
+* Add FutureChange<R> and StreamChange<R> which result is Future<R> or Stream<R>, Change<R> can have any object as result
+  
 [0.1.0] - beta release
 
 * Update CoreMwwmWidget. Change API